requirements.txt
Flask>=2.3
Flask-SQLAlchemy>=3.0
SQLAlchemy>=2.0
Flask-Login>=0.6
Flask-Bcrypt>=1.0
Flask-WTF>=1.1
Flask-Caching>=2.0
Flask-Migrate>=4.0
Authlib>=1.2
Pillow>=9.0
python-dotenv>=1.0
cryptography>=41.0
requests>=2.31
